// angular-integrations/code-grant-oidc-single-app-integration/projects/private-app/src/app/rest-client/model/bankAccountDetail.ts
/**
 * Swagger Customer - OpenAPI 3.0
 *
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccountStatusEnum } from './accountStatusEnum';
import { CurrencyEnum } from './currencyEnum';
import { AccountTypeEnum } from './accountTypeEnum';


export interface BankAccountDetail { 
    id?: string;
    sequenceNumber?: number;
    firstName?: string;
    lastName?: string;
    phone?: string;
    email?: string;
    street?: string;
    streetNumber?: string;
    postalCode?: string;
    city?: string;
    country?: string;
    accountNumber?: string;
    accountType?: AccountTypeEnum;
    ifscCode?: string;
    balance?: number;
    currency?: CurrencyEnum;
    status?: AccountStatusEnum;
}
export namespace BankAccountDetail {
}