// angular-integrations/code-grant-oidc-single-app-integration/projects/private-app/src/app/rest-client/model/bankAccountFilterResponse.ts
/**
 * Swagger Customer - OpenAPI 3.0
 *
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BankAccountDetail } from './bankAccountDetail';


export interface BankAccountFilterResponse { 
    data?: Array<BankAccountDetail>;
    foundItems?: number;
    totalItems?: number;
    previousPageToken?: number;
    nextPageToken?: number;
}